# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CISSP Exam Practice System",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/etl/datasets": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "List Datasets",
        "operationId": "list_datasets_api_etl_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/etl/datasets/{slug}": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "Get Dataset",
        "operationId": "get_dataset_api_etl_datasets__slug__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/runs": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_etl_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/etl/upload": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Upload Dataset",
        "description": "FR-IMP-01 / #35: upload a CSV/XLSX/JSON question file, materialize it as a\ndataset directory, and run a preview (extract+transform). Returns the same\nshape as ``POST /runs`` so the existing commit/rollback flow reuses verbatim.\n\nThe file is written to ``<etl_upload_root>/<dataset_slug>/questions.<ext>``\nso ``DatasetReader`` auto-detects it and re-reads at commit time for drift\ndetection (the same mechanism the JSONL path uses).",
        "operationId": "upload_dataset_api_etl_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_etl_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/etl/paste": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Paste Markdown",
        "description": "FR-IMP-02: interactive Markdown paste → materialize as questions.json → preview.",
        "operationId": "paste_markdown_api_etl_paste_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasteIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/etl/runs/{run_id}": {
      "get": {
        "tags": [
          "etl"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_etl_runs__run_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/runs/{run_id}/commit": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Commit Run",
        "operationId": "commit_run_api_etl_runs__run_id__commit_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/etl/runs/{run_id}/rollback": {
      "post": {
        "tags": [
          "etl"
        ],
        "summary": "Rollback Run",
        "operationId": "rollback_run_api_etl_runs__run_id__rollback_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_etl_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "dataset_slug": {
            "type": "string",
            "title": "Dataset Slug"
          }
        },
        "type": "object",
        "required": [
          "file",
          "dataset_slug"
        ],
        "title": "Body_upload_dataset_api_etl_upload_post"
      },
...
      "CreateRunIn": {
        "properties": {
          "dataset_slug": {
            "type": "string",
            "title": "Dataset Slug"
          }
        },
        "type": "object",
        "required": [
          "dataset_slug"
        ],
        "title": "CreateRunIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasteIn": {
        "properties": {
          "markdown": {
            "type": "string",
            "title": "Markdown"
          },
          "dataset_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Slug"
          }
        },
        "type": "object",
        "required": [
          "markdown"
        ],
        "title": "PasteIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}